+++ src/errors/GeneratingSchemaError.ts
@@ -1,9 +1,16 @@
 import { GraphQLError } from "graphql";
 
 export class GeneratingSchemaError extends Error {
-  constructor(public details: ReadonlyArray<GraphQLError>) {
-    super("Generating schema error");
+  details: ReadonlyArray<GraphQLError>;
 
+  constructor(details: ReadonlyArray<GraphQLError>) {
+    let errorMessage = "Some errors occurred while generating GraphQL schema:\n";
+    errorMessage += details.map(it => `  ${it.message}\n`);
+    errorMessage += "Please check the `details` property of the error to get more detailed info.";
+
+    super(errorMessage);
     Object.setPrototypeOf(this, new.target.prototype);
+
+    this.details = details;
   }
 }