+++ src/errors/WrongNullableListOptionError.ts
@@ -1,9 +1,13 @@
 import { NullableListOptions } from "../decorators/types";
 
 export class WrongNullableListOptionError extends Error {
-  constructor(typeOwnerName: string, nullable: boolean | NullableListOptions | undefined) {
+  constructor(
+    targetName: string,
+    propertyName: string,
+    nullable: boolean | NullableListOptions | undefined,
+  ) {
     super(
-      `Wrong nullable option set for ${typeOwnerName}. ` +
+      `Wrong nullable option set for ${targetName}#${propertyName}. ` +
         `You cannot combine non-list type with nullable '${nullable}'.`,
     );